Apache httpd 2.2 End-of-Life 2018-01-01
As previously announced, the Apache HTTP Server Project has discontinued all development and patch review of the 2.2.x series of releases.

The Apache HTTP Server Project had long committed to provide maintenance releases of the 2.2.x flavor through June of 2017, and to continue to publish some security source code patches until December of 2017. The final release 2.2.34 was published in July 2017, and no further evaluation of security risks will be published for 2.2.x releases.